Make sure we show we errored out while activating BE
This commit introduces changes to update the value of ret as by default that is set to SUCCESS, so in case this errors out - activation of a BE would have failed in reality but no error would have raised. It's hard to reproduce so it's possible there might be some other issues going on as well but this should at least improve the situation.