Do not run mid_activate hook when activating BE
This commit adds changes to not run mid_activate hook as we are seeing a case where the dataset in question is attempted to be mounted twice with the second try failing as desired. It should not be doing that but without a good reproduction case, we are not sure what exactly might be prompting this behavior. For now we disable the mid activate hook as it's not being used by the grub plugin which we are using.